pyproject.toml
[project]
name = "prag-v2"
version = "0.1.0"
description = "RAG system for academic research papers"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "qdrant-client>=1.7.0",
    "docling>=1.0.0",
    "pyyaml>=6.0",
    "python-multipart>=0.0.6",
    "httpx>=0.26.0",
    "ollama>=0.1.0",
    "fastembed>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
]
api = [
    "anthropic>=0.18.0",
    "google-generativeai>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
